refactor(client): split notification_section into parts
1482 LOC parent down to 438. Five sibling part files
(camera_preview, prefs_helpers, sound_picker, time_tile,
voice_video_section) keep their own private state via part-of
declarations sharing the parent's class scope. Behaviour-preserving;
settings tests still pass.
